Refactor browser trends component into modules
