+++ src/ui/views/browser/apps/trends.js
@@ -1,4 +1,4 @@
-import trendsApp from '../components/trends.js';
+import { render as renderTrendsComponent } from '../components/trends.js';
 import { getPageByType } from './pageLookup.js';
 
 export default function renderTrends(context = {}, model = {}) {
@@ -18,7 +18,7 @@ export default function renderTrends(context = {}, model = {}) {
   const mount = refs.body.querySelector('[data-role="trends-root"]');
   if (!mount) return null;
 
-  const summary = trendsApp.render(model, { mount, page });
+  const summary = renderTrendsComponent(model, { mount, page });
   const meta = summary?.meta || model?.highlights?.hot?.title || 'Trend insights ready';
   return { id: page.id, meta };
 }
